+++ RiskyClassicItemsSolution/Items/Common/MysteriousVial.cs
@@ -3,6 +3,8 @@ using R2API;
 using ClassicItemsReturns.Utils;
 using RoR2;
 using UnityEngine;
+using ClassicItemsReturns.Modules;
+using UnityEngine.AddressableAssets;
 
 namespace ClassicItemsReturns.Items.Common
 {//https://github.com/swuff-star/LostInTransit/blob/0fc3e096621a2ce65eef50f0e82db125c0730260/LIT/Assets/LostInTransit/Modules/Pickups/Items/MysteriousVial.cs
@@ -54,6 +56,32 @@ namespace ClassicItemsReturns.Items.Common
                 float levelFactor = 1f + 0.2f * (sender.level - 1f);
                 args.baseRegenAdd += ItemHelpers.StackingLinear(count, regen, regenPerStack) * levelFactor;
             }
+        }
+
+        protected override void CreateCraftableDef()
+        {
+            CraftableDef toInfusion = ScriptableObject.CreateInstance<CraftableDef>();
+            toInfusion.pickup = Addressables.LoadAssetAsync<ItemDef>("RoR2/Base/Infusion/Infusion.asset").WaitForCompletion();
+            toInfusion.recipes = new Recipe[]
+            {
+                new Recipe()
+                {
+                    amountToDrop = 1,
+                    ingredients = new RecipeIngredient[]
+                    {
+                        new RecipeIngredient()
+                        {
+                            pickup = ItemDef
+                        },
+                        new RecipeIngredient()
+                        {
+                            pickup = ItemDef
+                        }
+                    }
+                }
+            };
+            (toInfusion as ScriptableObject).name = "cdMysteriousVialToInfusion";
+            PluginContentPack.craftableDefs.Add(toInfusion);
         }
     }
 }